+++ simpleflow/settings/__init__.py
@@ -4,6 +4,5 @@ from future.utils import iteritems
 
 from . import base
 
-
 for k, v in iteritems(base.load()):
     setattr(sys.modules[__name__], k, v)