+++ simpleflow/workflow.py
@@ -1,7 +1,6 @@
 from __future__ import absolute_import
 
-
-from simpleflow.base import Submittable
+from simpleflow.base import Submittable, SubmittableContainer
 from simpleflow.signal import WaitForSignal
 from . import canvas
 from . import task
@@ -61,7 +60,7 @@ class Workflow(Submittable):
             return self._executor.submit(submittable, *args, **kwargs)
         elif isinstance(submittable, (task.Task, WaitForSignal)):
             return self._executor.submit(submittable)
-        elif isinstance(submittable, canvas.Group):
+        elif isinstance(submittable, SubmittableContainer):
             return submittable.submit(self._executor)
         else:
             raise TypeError('Bad type for {} activity ({})'.format(